% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdagwr.R
\name{fdagwr}
\alias{fdagwr}
\title{Functional Multi-Source Geographically Weighted Regression}
\description{
fdagwr provides the tools for fitting, depending on the non-stationary and, eventually, the number of sources of non-stationarity
a Functional Weighted Regression (FWR), a Functional Geographically Weighted Regression (FGWR), a Functional Mixed Geographically Weighted Regression (FMGWR)
or a Functional Multi-Source Geographically Weighted Regression (FMSGWR) model.
The package provides also the predictor for each one of the previous mentioned models.
Functional data, over a 1D domain, rather than response reconstruction weights, have to be represented as basis expansion in advance.
}
\references{
\itemize{
\item M. Fervari, A. Menafoglio and T. Bortolotti, \emph{"A Functional Multi-Source Geographically Weighted Regression for Ground Motion Modelling in Italy"}.2025.
\item T. Bortolotti, R. Peli, G. Lanzano, S. Sgobba and A. Menafoglio, \emph{"Weighted functional data analysis for the calibration of a ground motion model in italy"}. Journal of the American Statistical Association, 119(547):1697–1708, 2024.
\item L. Caramenti, A. Menafoglio, S. Sgobba and G. Lanzano, \emph{"Multi-source geographically weighted regression for regionalized ground-motion models"}. Spatial Statistics, 47:100610, 2022.
\item Source code: \href{https://github.com/AndreaEnricoFranzoni/FunctionalMultiSourceGeographicallyWeightedRegression}{fdagwr implementation}
}
}
\seealso{
\itemize{
\item Functional Multi-Source Geographically Weighted Regression ESC:
\itemize{
\item model fitting: \code{\link{FMSGWR_ESC}}
\item predictor: \code{\link{predict_FMSGWR_ESC}}}
\item Functional Multi-Source Geographically Weighted Regression SEC:
\itemize{
\item model fitting: \code{\link{FMSGWR_SEC}}
\item predictor: \code{\link{predict_FMSGWR_SEC}}}
\item Functional Mixed Geographically Weighted Regression:
\itemize{
\item model fitting: \code{\link{FMGWR}}
\item predictor: \code{\link{predict_FMGWR}}}
\item Functional Geographically Weighted Regression:
\itemize{
\item model fitting: \code{\link{FGWR}}
\item predictor: \code{\link{predict_FGWR}}}
\item Functional Weighted Regression:
\itemize{
\item model fitting: \code{\link{FWR}}
\item predictor: \code{\link{predict_FWR}}}}

#'@docType package
}
