% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fda_gwr_R_interface.R
\name{FMSGWR_ESC}
\alias{FMSGWR_ESC}
\title{FMSGWR_ESC}
\arguments{
\item{y_points}{\strong{\verb{numeric matrix}} of double containing the raw response: each row represents a specific abscissa for which the response evaluation is available, each column a statistical unit. Response is a already reconstructed.}

\item{t_points}{\strong{\verb{numeric vector}} of double with the abscissa points with respect of the raw evaluations of \p y_points are available (length of \p t_points is equal to the number of rows of \p y_points).}

\item{left_extreme_domain}{\strong{\code{double}} indicating the left extreme of the functional data domain (not necessarily the smaller element in \p t_points).}

\item{right_extreme_domain}{double indicating the right extreme of the functional data domain (not necessarily the biggest element in \p t_points).}

\item{coeff_y_points}{\strong{\verb{numeric matrix}} of double containing the coefficient of response's basis expansion: each row represents a specific basis (by default: B-spline) of the basis system used, each column a statistical unit.}

\item{knots_y_points}{\strong{\verb{numeric vector}} of double with the abscissa points with respect which the basis expansions of the response and response reconstruction weights are performed (all elements contained in \link{a,b}).}

\item{degree_basis_y_points}{\strong{\verb{non-negative integer}}: the degree of the basis used for the basis expansion of the (functional) response. Default explained below (can be \strong{\code{NULL}}).}

\item{n_basis_y_points}{\strong{\verb{positive integer}}: number of basis for the basis expansion of the (functional) response. It must match number of rows of coeff_y_points. Default explained below (can be \strong{\code{NULL}}).}

\item{coeff_rec_weights_y_points}{\strong{\verb{numeric matrix}} of double containing the coefficients of the basis expansion of the weights to reconstruct the (functional) response: each row represents a specific basis (by default: B-spline) of the basis system used, each column a statistical unit.}

\item{degree_basis_rec_weights_y_points}{\strong{\verb{non-negative integer}}: the degree of the basis used for response reconstruction weights. Default explained below (can be \strong{\code{NULL}}).}

\item{n_basis_rec_weights_y_points}{\strong{\verb{positive integer}}: number of basis for the basis expansion of response reconstruction weights. It must match number of rows of \p coeff_rec_weights_y_points. Default explained below (can be \strong{\code{NULL}}).}

\item{coeff_stationary_cov}{\strong{\verb{list of numeric matrices}} of double: element i-th containing the coefficients for the basis expansion of the i-th stationary covariate: each row represents a specific basis (by default: B-spline) of the basis system used, each column a statistical unit.
The name of the i-th element is the name of the i-th stationary covariate (default: \strong{\code{"reg.Ci"}} if no name present).}

\item{basis_types_stationary_cov}{\strong{\verb{vector of strings}}, element i-th containing the type of basis used for the i-th stationary covariate basis expansion. Possible values: \strong{\code{"bsplines"}}, \strong{\code{"constant"}}. Defalut: \strong{\code{"bsplines"}}.}

\item{knots_stationary_cov}{\strong{\verb{numeric vector}} of double with the abscissa points with respect which the basis expansions of the stationary covariates are performed (all elements contained in \link{a,b}).}

\item{degrees_basis_stationary_cov}{\strong{\verb{integer vector}} of non-negative integers: element i-th is the degree of the basis used for the basis expansion of the i-th stationary covariate. Default explained below (can be \strong{\code{NULL}})).}

\item{n_basis_stationary_cov}{\strong{\verb{integer vector}} of positive integers: element i-th is the number of basis for the basis expansion of the i-th stationary covariate. It must match number of rows of the i-th element of \p coeff_stationary_cov. Default explained below (can be \strong{\code{NULL}}).}

\item{penalization_stationary_cov}{\strong{\verb{numeric vector}} of non-negative double: element i-th is the penalization used for the i-th stationary covariate.}

\item{knots_beta_stationary_cov}{\strong{\verb{numeric vector}} of double with the abscissa points with respect which the basis expansions of the stationary covariates functional regression coefficients are performed (all elements contained in \link{a,b}).}

\item{degrees_basis_beta_stationary_cov}{\strong{\verb{integer vector}}  of non-negative integers: element i-th is the degree of the basis used for the basis expansion of the i-th stationary covariate functional regression coefficients. Default explained below (can be \strong{\code{NULL}}).}

\item{n_basis_beta_stationary_cov}{\strong{\verb{integer vector}} of positive integers: element i-th is the number of basis for the basis expansion of the i-th stationary covariate functional regression coefficients. Default explained below (can be \strong{\code{NULL}}).}

\item{coeff_events_cov}{\strong{\verb{list of numeric matrices}} of double: element i-th containing the coefficients for the basis expansion of the i-th events-dependent covariate: each row represents a specific basis (by default: B-spline) of the basis system used, each column a statistical unit.
The name of the i-th element is the name of the i-th events-dependent covariate (default: \strong{\code{"reg.Ei"}} if no name present).}

\item{basis_types_events_cov}{\strong{\verb{vector of strings}}, element i-th containing the type of basis used for the i-th events-dependent covariate basis expansion. Possible values: \strong{\code{"bsplines"}}, \strong{\code{"constant"}}. Defalut: \strong{\code{"bsplines"}}.}

\item{knots_events_cov}{\strong{\verb{numeric vector}} of double with the abscissa points with respect which the basis expansions of the events-dependent covariates are performed (all elements contained in \link{a,b}).}

\item{degrees_basis_events_cov}{\strong{\verb{integer vector}} of non-negative integers: element i-th is the degree of the basis used for the basis expansion of the i-th events-dependent covariate. Default explained below (can be \strong{\code{NULL}}).}

\item{n_basis_events_cov}{\strong{\verb{integer vector}} of positive integers: element i-th is the number of basis for the basis expansion of the i-th events-dependent covariate. It must match number of rows of the i-th element of \p coeff_events_cov. Default explained below (can be \strong{\code{NULL}}).}

\item{penalization_events_cov}{\strong{\verb{numeric vector}} of non-negative double: element i-th is the penalization used for the i-th events-dependent covariate.}

\item{coordinates_events}{\strong{\verb{numeric matrix}} of double containing the UTM coordinates of the event of each statistical unit: each row represents a statistical unit, each column a coordinate (2 columns).}

\item{kernel_bandwith_events}{\strong{\verb{positive double}} indicating the bandwith of the gaussian kernel used to smooth the distances within events.}

\item{knots_beta_events_cov}{\strong{\verb{numeric vector}} of double with the abscissa points with respect which the basis expansions of the events-dependent covariates functional regression coefficients are performed (all elements contained in \link{a,b}).}

\item{degrees_basis_beta_events_cov}{\strong{\verb{integer vector}} of non-negative integers: element i-th is the degree of the basis used for the basis expansion of the i-th events-dependent covariate functional regression coefficient. Default explained below (can be \strong{\code{NULL}}).}

\item{n_basis_beta_events_cov}{\strong{\verb{integer vector}} of positive integers: element i-th is the number of basis for the basis expansion of the i-th events-dependent covariate functional regression coefficient. Default explained below (can be \strong{\code{NULL}}).}

\item{coeff_stations_cov}{\strong{\verb{list of numeric matrices}} of double: element i-th containing the coefficients for the basis expansion of the i-th stations-dependent covariate: each row represents a specific basis (by default: B-spline) of the basis system used, each column a statistical unit.
The name of the i-th element is the name of the i-th stations-dependent covariate (default: \strong{\code{"reg.Si"}} ).}

\item{basis_types_stations_cov}{\strong{\verb{vector of strings}}, element i-th containing the type of basis used for the i-th stations-dependent covariates basis expansion. Possible values: \strong{\code{"bsplines"}}, \strong{\code{"constant"}}. Defalut: \strong{\code{"bsplines"}}.}

\item{knots_stations_cov}{\strong{\verb{numeric vector}} of double with the abscissa points with respect which the basis expansions of the stations-dependent covariates are performed (all elements contained in \link{a,b}).}

\item{degrees_basis_stations_cov}{\strong{\verb{integer vector}} of non-negative integers: element i-th is the degree of the basis used for the basis expansion of the i-th stations-dependent covariate. Default explained below (can be \strong{\code{NULL}}).}

\item{n_basis_stations_cov}{\strong{\verb{integer vector}} of positive integers: element i-th is the number of basis for the basis expansion of the i-th stations-dependent covariate. It must match number of rows of the i-th element of \p coeff_stations_cov. Default explained below (can be \strong{\code{NULL}}).}

\item{penalization_stations_cov}{\strong{\verb{numeric vector}} of non-negative double: element i-th is the penalization used for the i-th stations-dependent covariate.}

\item{coordinates_stations}{\strong{\verb{numeric matrix}} of double containing the UTM coordinates of the station of each statistical unit: each row represents a statistical unit, each column a coordinate (2 columns).}

\item{kernel_bandwith_stations}{\strong{\verb{positive double}} indicating the bandwith of the gaussian kernel used to smooth the distances within stations.}

\item{knots_beta_stations_cov}{\strong{\verb{numeric vector}} of double with the abscissa points with respect which the basis expansions of the stations-dependent covariates functional regression coefficients are performed (all elements contained in \link{a,b}).}

\item{degrees_basis_beta_stations_cov}{\strong{\verb{integer vector}} of non-negative integers: element i-th is the degree of the basis used for the basis expansion of the i-th stations-dependent covariate functional regression coefficient. Default explained below (can be \strong{\code{NULL}}).}

\item{n_basis_beta_stations_cov}{\strong{\verb{integer vector}} of positive integers: element i-th is the number of basis for the basis expansion of the i-th stations-dependent covariate functional regression coefficient. Default explained below (can be \strong{\code{NULL}}).}

\item{in_cascade_estimation}{\strong{\code{bool}}: if false, an exact algorithm taking account for the interaction within non-stationary covariates is used to fit the model. Otherwise, the model is fitted in cascade. The first option is more precise, but way more computationally intensive.}

\item{n_knots_smoothing}{\strong{\verb{positive integer}}: number of knots used to perform the smoothing on the response obtained leaving out all the non-stationary components (default: \strong{\code{100}}).}

\item{n_intervals_quadrature}{\strong{\verb{positive integer}}: number of intervals used while performing integration via midpoint (rectangles) quadrature rule (default: \strong{\code{100}}).}

\item{num_threads}{\strong{\verb{positive integer}}: number of threads to be used in OMP parallel directives. Default: \strong{\code{NULL}}, that is equivalent to the maximum number of cores available in the machine.}

\item{basis_type_y_points}{\strong{\code{string}} containing the type of basis used for the functional response basis expansion. Possible values: \strong{\code{"bsplines"}}, \strong{\code{"constant"}}. Defalut: \strong{\code{"bsplines"}}.}

\item{basis_type_rec_weights_y_points}{\strong{\code{string}} containing the type of basis used for the weights to reconstruct the functional response basis expansion. Possible values: \strong{\code{"bsplines"}}, \strong{\code{"constant"}}. Defalut: \strong{\code{"bsplines"}}.}

\item{basis_types_beta_stationary_cov}{\strong{\verb{vector of strings}}, element i-th containing the type of basis used for the i-th stationary covariate functional regression coefficients basis expansion. Possible values: \strong{\code{"bsplines"}}, \strong{\code{"constant"}}. Defalut: \strong{\code{"bsplines"}}.}

\item{basis_types_beta_events_cov}{\strong{\verb{vector of strings}}, element i-th containing the type of basis used for the i-th events-dependent covariate functional regression coefficients basis expansion. Possible values: \strong{\code{"bsplines"}}, \strong{\code{"constant"}}. Defalut: \strong{\code{"bsplines"}}.}

\item{basis_types_beta_stations_cov}{\strong{\verb{vector of strings}}, element i-th containing the type of basis used for the i-th stations-dependent covariate functional regression coefficients basis expansion. Possible values: \strong{\code{"bsplines"}}, \strong{\code{"constant"}}. Defalut: \strong{\code{"bsplines"}}.}
}
\value{
\strong{\code{list}} containing:
\itemize{
\item 'FGWR': \strong{\code{string}}: the type of fgwr used (\strong{\code{"FMSGWR_ESC"}});
\item 'EstimationTechnique': \strong{\code{string}}: \strong{\code{"Exact"}} if \p in_cascade_estimation false, \strong{\code{"Cascade"}} if \p in_cascade_estimation true
\item 'Bc': \strong{\code{list}} containing, for each stationary covariate regression coefficient (each element is named with the element names in the list \p coeff_stationary_cov (default, if not given: \strong{\code{"CovC*"}})):
\itemize{
\item 'basis_coeff': \strong{\verb{numeric vector}} of double, dimension Lc_jx1,containing the coefficients of the basis expansion of the beta.
\item 'basis_type': \strong{\code{string}} containing the basis type over which the beta basis expansion is performed. Possible values: \strong{\code{"bsplines"}}, \strong{\code{"constant"}}. (Respective element of \p basis_types_beta_stationary_cov).
\item 'basis_num': \strong{\verb{positive integer}}: the number of basis used for performing the beta basis expansion (respective elements of \p n_basis_beta_stationary_cov).
\item 'knots': \strong{\verb{numeric vector}} of double: knots used to create the basis system for the beta (it is the input \p knots_beta_stationary_cov).
}
\item 'Beta_c': \strong{\code{list}} containing, for each stationary covariate regression coefficient (each element is named with the element names in the list coeff_stationary_cov (default, if not given: \strong{\code{"CovC*"}})) a list with:
\itemize{
\item 'Beta_eval': \strong{\verb{numeric vector}} of double containing the discrete evaluations of the stationary beta.
\item 'Abscissa': \strong{\verb{numeric vector}} of double containing the domain points for which the evaluation of the beta is available (it is the input \p t_points).
}
\item 'Be': \strong{\code{list}} containing, for each event-dependent covariate regression coefficient (each element is named with the element names in the list \p coeff_events_cov (default, if not given: \strong{\code{"CovE*"}})):
\itemize{
\item 'basis_coeff': \strong{\code{list}}, containing, for each statistical unit, a \strong{\verb{numeric vector}} of double, dimension Le_jx1,containing the coefficients of the basis expansion of the beta for that unit.
\item 'basis_type': \strong{\code{string}} containing the basis type over which the beta basis expansion is performed. Possible values: \strong{\code{"bsplines"}}, \strong{\code{"constant"}}. (Respective element of \p basis_types_beta_events_cov).
\item 'basis_num': \strong{\verb{positive integer}}: the number of basis used for performing the beta basis expansion (respective elements of \p n_basis_beta_events_cov).
\item 'knots': \strong{\verb{numeric vector}} of double: knots used to create the basis system for the beta (it is the input \p knots_beta_events_cov).
}
\item 'Beta_e': \strong{\code{list}} containing, for each event-dependent covariate regression coefficient (each element is named with the element names in the list \p coeff_events_cov (default, if not given: \strong{\code{"CovE*"}})) a list with:
\itemize{
\item 'Beta_eval': \strong{\code{list}}, containing, for each statistical unit, a \strong{\verb{numeric vector}} of double containing the discrete evaluations of the non-stationary beta, one for each unit.
\item 'Abscissa': \strong{\verb{numeric vector}} of double containing the domain points for which the evaluation of the beta is available (it is the input \p t_points).
}
\item 'Bs': \strong{\code{list}} containing, for each station-dependent covariate regression coefficient (each element is named with the element names in the list \p coeff_stations_cov (default, if not given: \strong{\code{"CovS*"}})):
\itemize{
\item 'basis_coeff': \strong{\code{list}}, containing, for each statistical unit, a \strong{\verb{numeric vector}} of double, dimension Ls_jx1,containing the coefficients of the basis expansion of the beta for that unit.
\item 'basis_type': \strong{\code{string}} containing the basis type over which the beta basis expansion is performed. Possible values: \strong{\code{"bsplines"}}, \strong{\code{"constant"}}. (Respective element of \p basis_types_beta_stations_cov).
\item 'basis_num': \strong{\verb{positive integer}}: the number of basis used for performing the beta basis expansion (respective elements of \p n_basis_beta_stations_cov).
\item 'knots': \strong{\verb{numeric vector}} of double: knots used to create the basis system for the beta (it is the input \p knots_beta_stations_cov).
}
\item 'Beta_e': \strong{\code{list}} containing, for each station-dependent covariate regression coefficient (each element is named with the element names in the list \p coeff_stations_cov (default, if not given: \strong{\code{"CovS*"}})) a list with:
\itemize{
\item 'Beta_eval': \strong{\code{list}}, containing, for each statistical unit, a \strong{\verb{numeric vector}} of double containing the discrete evaluations of the non-stationary beta, one for each unit.
\item 'Abscissa': \strong{\verb{numeric vector}} of double containing the domain points for which the evaluation of the beta is available (it is the input \p t_points).
}
\item 'predictor_info': \strong{\code{list}} containing partial residuals and information of the fitted model to perform predictions for new statistical units:
\itemize{
\item 'partial_res': \strong{\code{list}} containing information to compute the partial residuals:
\itemize{
\item 'c_tilde_hat': \strong{\verb{numeric vector}} of double with the basis expansion coefficients of the response minus the stationary component of the phenomenon (if \p in_cascade_estimation is true, contains only 0s).
\item 'A__': \strong{\verb{list of numeric matrices}}, containing the operator A_e for each statistical unit (if \p in_cascade_estimation is true, each matrix contains only 0s).
\item 'B__for_K': \strong{\verb{list of numeric matrices}}, containing the operator B_e used for the K_e_s(t) computation, for each statistical unit (if \p in_cascade_estimation is true, each matrix contains only 0s).
}
\item 'inputs_info': \strong{\code{list}} containing information about the data used to fit the model:
\itemize{
\item 'Response': \strong{\code{list}}:
\itemize{
\item 'basis_num': \strong{\verb{positive integer}}: number of basis used to make the basis expansion of the functional response (element \p n_basis_y_points).
\item 'basis_type': \strong{\code{string}}: basis used to make the basis expansion of the functional response. Possible values: \strong{\code{"bsplines"}}, \strong{\code{"constant"}} (element \p basis_type_y_points).
\item 'basis_deg': \strong{\verb{positive integer}}: degree of basis used to make the basis expansion of the functional response (element \p degree_basis_y_points).
\item 'knots': \strong{\verb{numeric vector}}: knots used to make the basis expansion of the functional response (element \p knots_y_points).
\item 'basis_coeff': \strong{\verb{numeric matrix}}: coefficients of the basis expansion of the functional response (element \p coeff_y_points).
}
\item 'ResponseReconstructionWeights': \strong{\code{list}}:
\itemize{
\item 'basis_num': \strong{\verb{positive integer}}: number of basis used to make the basis expansion of the functional response (element \p n_basis_rec_weights_y_points).
\item 'basis_type': \strong{\code{string}}: basis used to make the basis expansion of the functional response. Possible values: \strong{\code{"bsplines"}}, \strong{\code{"constant"}} (element \p basis_type_rec_weights_y_points).
\item 'basis_deg': \strong{\verb{positive integer}}: degree of basis used to make the basis expansion of the functional response (element \p degree_basis_rec_weights_y_points).
\item 'knots': \strong{\verb{numeric vector}}: knots used to make the basis expansion of the functional response (element \p knots_y_points).
\item 'basis_coeff': \strong{\verb{numeric matrix}}: coefficients of the basis expansion of the functional response (element \p coeff_rec_weights_y_points).
}
\item 'cov_Stationary': \strong{\code{list}}:
\itemize{
\item 'number_covariates': \strong{\verb{positive integer}}: number of stationary covariates (length of \p coeff_stationary_cov).
\item 'basis_num': \strong{\verb{integer vector}} of positive integer: numbers of basis used to make the basis expansion of the functional stationary covariates (respective elements of \p n_basis_stationary_cov).
\item 'basis_type': \strong{\verb{vector of strings}}: types of basis used to make the basis expansion of the functional stationary covariates. Possible values: \strong{\code{"bsplines"}}, \strong{\code{"constant"}} (respective elements of \p basis_types_stationary_cov).
\item 'basis_deg': \strong{\verb{integer vector}} of positive integers: degrees of basis used to make the basis expansion of the functional stationary covariates (respective elements of \p degrees_basis_stationary_cov).
\item 'knots': \strong{\verb{numeric vector}}: knots used to make the basis expansion of the functional stationary covariates (respective elements of \p knots_stationary_cov).
\item 'basis_coeff': \strong{\verb{list of numeric matrices}}: coefficients of the basis expansion of the functional stationary covariates  (respective elements of \p coeff_stationary_cov).
}
\item 'beta_Stationary': \strong{\code{list}}:
\itemize{
\item 'basis_num': \strong{\verb{integer vector}} of positive integer: numbers of basis used to make the basis expansion of the functional regression coefficients of the stationary covariates (respective elements of \p n_basis_beta_stationary_cov).
\item 'basis_type': \strong{\verb{vector of strings}}: types of basis used to make the basis expansion of the functional regression coefficients of the stationary covariates. Possible values: \strong{\code{"bsplines"}}, \strong{\code{"constant"}} (respective elements of \p basis_types_beta_stationary_cov).
\item 'basis_deg': \strong{\verb{integer vector}} of positive integers: degrees of basis used to make the basis expansion of the functional regression coefficients of the stationary covariates (respective elements \p degrees_basis_beta_stationary_cov).
\item 'knots': knots used to make the basis expansion of the functional regression coefficients of the stationary covariates (element \p knots_beta_stationary_cov).
}
\item 'cov_Event': \strong{\code{list}}:
\itemize{
\item 'number_covariates': \strong{\verb{positive integer}}: number of event-dependent covariates (length of \p coeff_events_cov).
\item 'basis_num': \strong{\verb{integer vector}} of positive integer: numbers of basis used to make the basis expansion of the functional event-dependent covariates (respective elements of \p n_basis_events_cov).
\item 'basis_type': \strong{\verb{vector of strings}}: types of basis used to make the basis expansion of the functional event-dependent covariates. Possible values: \strong{\code{"bsplines"}}, \strong{\code{"constant"}} (respective elements of \p basis_types_events_cov).
\item 'basis_deg': \strong{\verb{integer vector}} of positive integers: degrees of basis used to make the basis expansion of the functional event-dependent covariates (respective elements \p degrees_basis_events_cov).
\item 'knots': \strong{\verb{numeric vector}}: knots used to make the basis expansion of the functional event-dependent covariates (respective elements of \p knots_events_cov).
\item 'basis_coeff': \strong{\verb{list of numeric matrices}}: coefficients of the basis expansion of the functional event-dependent covariates  (respective elements of \p coeff_events_cov).
\item 'penalizations': \strong{\verb{numeric vector}} of positive double: penalizations of the event-dependent covariates (respective elements of \p penalization_events_cov)
\item 'coordinates': \strong{\verb{numeric matrix}}: UTM coordinates of the events of the training data (element \p coordinates_events).
\item 'kernel_bwd': \strong{\code{double}}: bandwith of the gaussian kernel used to smooth distances of the events (element \p kernel_bandwith_events).
}
\item 'beta_Event': \strong{\code{list}}:
\itemize{
\item 'basis_num': \strong{\verb{integer vector}} of positive integer: numbers of basis used to make the basis expansion of the functional regression coefficients of the event-dependent covariates (respective elements of \p n_basis_beta_events_cov).
\item 'basis_type': \strong{\verb{vector of strings}}: types of basis used to make the basis expansion of the functional regression coefficients of the event-dependent covariates. Possible values: \strong{\code{"bsplines"}}, \strong{\code{"constant"}} (respective elements of \p basis_types_beta_events_cov).
\item 'basis_deg': \strong{\verb{integer vector}} of positive integers: degrees of basis used to make the basis expansion of the functional regression coefficients of the event-dependent covariates (element \p degrees_basis_beta_events_cov).
\item 'knots': knots used to make the basis expansion of the functional regression coefficients of the event-dependent covariates (element \p knots_beta_events_cov).
}
\item 'cov_Station': \strong{\code{list}}:
\itemize{
\item 'number_covariates': \strong{\verb{positive integer}}: number of station-dependent covariates (length of \p coeff_stations_cov).
\item 'basis_num': \strong{\verb{integer vector}} of positive integer: numbers of basis used to make the basis expansion of the functional station-dependent covariates (respective elements of \p n_basis_stations_cov).
\item 'basis_type': \strong{\verb{vector of strings}}: types of basis used to make the basis expansion of the functional station-dependent covariates. Possible values: \strong{\code{"bsplines"}}, \strong{\code{"constant"}} (respective elements of \p basis_types_stations_cov).
\item 'basis_deg': \strong{\verb{integer vector}} of positive integers: degrees of basis used to make the basis expansion of the functional station-dependent covariates (respective elements \p degrees_basis_stations_cov).
\item 'knots': \strong{\verb{numeric vector}}: knots used to make the basis expansion of the functional station-dependent covariates (respective elements of \p knots_stations_cov).
\item 'basis_coeff': \strong{\verb{list of numeric matrices}}: coefficients of the basis expansion of the functional station-dependent covariates  (respective elements of \p coeff_stations_cov).
\item 'penalizations': \strong{\verb{numeric vector}} of positive double: penalizations of the station-dependent covariates (respective elements of \p penalization_stations_cov)
\item 'coordinates': \strong{\verb{numeric matrix}}: UTM coordinates of the stations of the training data (element \p coordinates_stations).
\item 'kernel_bwd': \strong{\code{double}}: bandwith of the gaussian kernel used to smooth distances of the stations (element \p kernel_bandwith_stations).
}
\item 'beta_Station': \strong{\code{list}}:
\itemize{
\item 'basis_num': \strong{\verb{integer vector}} of positive integer: numbers of basis used to make the basis expansion of the functional regression coefficients of the station-dependent covariates (respective elements of \p n_basis_beta_stations_cov).
\item 'basis_type': \strong{\verb{vector of strings}}: types of basis used to make the basis expansion of the functional regression coefficients of the station-dependent covariates. Possible values: \strong{\code{"bsplines"}}, \strong{\code{"constant"}} (respective elements of \p basis_types_beta_stations_cov).
\item 'basis_deg': \strong{\verb{integer vector}} of positive integers: degrees of basis used to make the basis expansion of the functional regression coefficients of the station-dependent covariates (element \p degrees_basis_beta_stations_cov).
\item 'knots': knots used to make the basis expansion of the functional regression coefficients of the station-dependent covariates (element \p knots_beta_stations_cov).
}
}
\item 'a': \strong{\code{double}}: domain left extreme  (element \p left_extreme_domain).
\item 'b': \strong{\code{double}}: domain right extreme  (element \p right_extreme_domain).
\item 'abscissa': \strong{\verb{numeric vector}} of double: abscissa for which the evaluations of the functional data are available (element \p t_points).
\item 'InCascadeEstimation': element \p in_cascade_estimation.
}
}
}
\description{
Fitting a Functional Multi-Source Geographically Weighted Regression ESC model. The covariates are functional objects, divided into
three categories: stationary covariates (C), constant over geographical space, event-dependent covariates (E), that vary depending on the spatial coordinates of the event,
station-dependent covariates (S), that vary depending on the spatial coordinates of the stations that measure the event. Regression coefficients are estimated
in the following order: C, S, E. The functional response is already reconstructed according to the method proposed by Bortolotti et Al. (2024)
}
\details{
Constant basis are used, for a covariate, if it resembles a scalar shape. It consists of a straight line with y-value equal to 1 all over the data domain.
Can be seen as a B-spline basis with degree 0, number of basis 1, using one knot, consequently having only one coefficient for the only basis for each statistical unit.
fdagwr sets all the feats accordingly if reads constant basis.
However, recall that the response is a functional datum, as the regressors coefficients. Since the package's basis variety could be hopefully enlarged in the future
(for example, introducing Fourier basis for handling data that present periodical behaviors), the input parameters regarding basis types for response, response reconstruction
weights and regressors coefficients are left at the end of the input list, and defaulted as \strong{\code{NULL}}. Consequently they will use a B-spline basis system, and should NOT use a constant basis,
Recall to perform externally the basis expansion before using the package, and afterwards passing basis types, degree and number and basis expansion coefficients and knots coherently
}
\note{
A little excursion about degree and number of basis passed as input. For each specific covariate, or the response, if using B-spline basis, remember that number of knots = number of basis - degree + 1.
By default, if passing \strong{\code{NULL}}, fdagwr uses a cubic B-spline system of basis, the number of basis is computed coherently from the number of knots (that is the only mandatory input parameter).
Passing only the degree of the bsplines, the number of basis used will be set accordingly, and viceversa if passing only the number of basis.
But, take care that the number of basis used has to match the number of rows of coefficients matrix (for EACH type of basis). If not, an exception is thrown. No problems arise if letting fdagwr defaulting the number of basis.
For response and response reconstruction weights, degree and number of basis consist of integer, and can be \strong{\code{NULL}} For all the regressors, and their coefficients, the inputs consist of vector of integers:
if willing to pass a default parameter, all the vector has to be defaulted (if passing \strong{\code{NULL}}, a vector with all 3 for the degrees is passed, for example)
}
\references{
\itemize{
\item Paper: \href{https://www.researchgate.net/publication/377251714_Weighted_Functional_Data_Analysis_for_the_Calibration_of_a_Ground_Motion_Model_in_Italy}{Functional Response Reconstruction Weights}
\item Source code: \href{https://github.com/AndreaEnricoFranzoni/FunctionalMultiSourceGeographicallyWeightedRegression}{fdagwr implementation}
}
}
\author{
Andrea Enrico Franzoni
}
