% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fda_gwr_R_interface.R
\name{predict_FWR}
\alias{predict_FWR}
\title{predict_FWR}
\arguments{
\item{coeff_stationary_cov_to_pred}{\strong{\verb{list of numeric matrices}} of double: element i-th containing the coefficients for the basis expansion of the i-th stationary covariate to be predicted: each row represents a specific basis (by default: B-spline) of the basis system used, each column a statistical unit to be predicted.}

\item{units_to_be_predicted}{\strong{\verb{positive integer}}: number of units to be predicted}

\item{abscissa_ev}{\strong{\verb{numeric vector}} of double: abscissa for which then evaluating the predicted response and betas, stationary and non-stationary, which have to be recomputed}

\item{model_fitted}{\strong{\code{list}} containing:
\itemize{
\item 'FGWR': \strong{\code{string}}: the type of fwr used (\strong{\code{"FWR"}});
\item 'Bc': \strong{\code{list}} containing, for each stationary covariate regression coefficient (each element is named with the element names in the list coeff_stationary_cov (default, if not given: \strong{\code{"CovC*"}})):
\itemize{
\item 'basis_coeff': \strong{\verb{numeric vector}} of double, dimension Lc_jx1,containing the coefficients of the basis expansion of the beta.
\item 'basis_type': \strong{\code{string}} containing the basis type over which the beta basis expansion is performed. Possible values: \strong{\code{"bsplines"}}, \strong{\code{"constant"}}. (Respective element of basis_types_beta_stationary_cov).
\item 'basis_num': \strong{\verb{positive integer}}: the number of basis used for performing the beta basis expansion (respective elements of n_basis_beta_stationary_cov).
\item 'knots': \strong{\verb{numeric vector}} of double: knots used to create the basis system for the beta (it is the input knots_beta_stationary_cov).
}
\item 'Beta_c': \strong{\code{list}} containing, for each stationary covariate regression coefficient (each element is named with the element names in the list coeff_stationary_cov (default, if not given: \strong{\code{"CovC*"}})) a list with:
\itemize{
\item 'Beta_eval': \strong{\verb{numeric vector}} of double containing the discrete evaluations of the stationary beta.
\item 'Abscissa': \strong{\verb{numeric vector}} of double containing the domain points for which the evaluation of the beta is available (it is the input t_points).
}
\item 'predictor_info': \strong{\code{list}} containing partial residuals and information of the fitted model to perform predictions for new statistical units:
\itemize{
\item 'inputs_info': \strong{\code{list}} containing information about the data used to fit the model:
\itemize{
\item 'Response': \strong{\code{list}}:
\itemize{
\item 'basis_num': \strong{\verb{positive integer}}: number of basis used to make the basis expansion of the functional response (element n_basis_y_points).
\item 'basis_type': \strong{\code{string}}: basis used to make the basis expansion of the functional response. Possible values: \strong{\code{"bsplines"}}, \strong{\code{"constant"}} (element basis_type_y_points).
\item 'basis_deg': \strong{\verb{positive integer}}: degree of basis used to make the basis expansion of the functional response (element degree_basis_y_points).
\item 'knots': \strong{\verb{numeric vector}}: knots used to make the basis expansion of the functional response (element knots_y_points).
}
\item 'cov_Stationary': \strong{\code{list}}:
\itemize{
\item 'number_covariates': \strong{\verb{positive integer}}: number of stationary covariates (length of coeff_stationary_cov).
\item 'basis_num': \strong{\verb{integer vector}} of positive integer: numbers of basis used to make the basis expansion of the functional stationary covariates (respective elements of n_basis_stationary_cov).
\item 'basis_type': \strong{\verb{vector of strings}}: types of basis used to make the basis expansion of the functional stationary covariates. Possible values: \strong{\code{"bsplines"}}, \strong{\code{"constant"}} (respective elements of basis_types_stationary_cov).
\item 'basis_deg': \strong{\verb{integer vector}} of positive integers: degrees of basis used to make the basis expansion of the functional stationary covariates (respective elements of degrees_basis_stationary_cov).
\item 'knots': \strong{\verb{numeric vector}}: knots used to make the basis expansion of the functional stationary covariates (respective elements of knots_stationary_cov).
}
\item 'beta_Stationary': \strong{\code{list}}:
\itemize{
\item 'basis_num': \strong{\verb{integer vector}} of positive integer: numbers of basis used to make the basis expansion of the functional regression coefficients of the stationary covariates (respective elements of n_basis_beta_stationary_cov).
\item 'basis_type': \strong{\verb{vector of strings}}: types of basis used to make the basis expansion of the functional regression coefficients of the stationary covariates. Possible values: \strong{\code{"bsplines"}}, \strong{\code{"constant"}} (respective elements of basis_types_beta_stationary_cov).
\item 'basis_deg': \strong{\verb{integer vector}} of positive integers: degrees of basis used to make the basis expansion of the functional regression coefficients of the stationary covariates (respective elements degrees_basis_beta_stationary_cov).
\item 'knots': \strong{\verb{numeric vector}}: knots used to make the basis expansion of the functional regression coefficients of the stationary covariates (element knots_beta_stationary_cov).
}
}
\item 'a': \strong{\code{double}}: domain left extreme  (element left_extreme_domain).
\item 'b': \strong{\code{double}}: domain right extreme  (element right_extreme_domain).
\item 'abscissa': \strong{\verb{numeric vector}} of double: abscissa for which the evaluations of the functional data are available (element t_points).
}
}}

\item{n_knots_smoothing_pred}{\strong{\verb{positive integer}}: number of knots used to smooth predicted response and non-stationary, obtaining basis expansion coefficients with respect to the training basis (default: \strong{\code{100}}).}

\item{num_threads}{\strong{\verb{positive integer}}: number of threads to be used in OMP parallel directives. Default: \strong{\code{NULL}}, that is equivalent to the maximum number of cores available in the machine.}
}
\value{
\strong{\code{list}} containing:
\itemize{
\item 'FGWR_predictor': \strong{\code{string}}: model used to predict (\strong{\code{"predictor_FWR"}}).
\item 'prediction': \strong{\code{list}} containing:
\itemize{
\item 'evaluation': \strong{\code{list}} containing the evaluation of the prediction:
\itemize{
\item 'prediction_ev': \strong{\code{list}} containing, for each unit to be predicted, the raw evaluations of the predicted response.
\item 'abscissa_ev': \strong{\verb{numeric vector}} containing the abscissa points for which the prediction evaluation is available (element abscissa_ev).
}
\item 'fd': \strong{\code{list}} containing the prediction functional description:
\itemize{
\item 'prediction_basis_coeff': \strong{\verb{numeric matrix}} containing the prediction basis expansion coefficients (each row a basis, each column a new statistical unit).
\item 'prediction_basis_type': \strong{\code{string}}: basis type used for the predicted response basis expansion (from model_fitted).
\item 'prediction_basis_num': \strong{\verb{positive integer}}: number of basis used for the predicted response basis expansion (from model_fitted).
\item 'prediction_basis_deg': \strong{\verb{positive integer}}: degree of basis used for the predicted response basis expansion (from model_fitted)
\item 'prediction_knots': \strong{\verb{numeric vector}}: knots used for the predicted response smoothing (n_knots_smoothing_pred equally spaced knots in the functional datum domain)
}
}
\item 'Bc_pred': \strong{\code{list}} containing, for each stationary covariate:
\itemize{
\item 'basis_coeff': \strong{\verb{numeric matrix}} of double containing the fitted basis expansion coefficients of the beta (from model_fitted).
\item 'basis_num': \strong{\verb{positive integer}}: number of basis used for the beta basis expansion (from model_fitted).
\item 'basis_type': \strong{\code{string}}: type of basis used for the beta basis expansion (from model_fitted).
\item 'knots': \strong{\verb{numeric vector}}: knots used for the beta basis expansion (from model_fitted).
}
\item 'Beta_c_pred': \strong{\code{list}} containing, for each stationary covariate:
\itemize{
\item 'Beta_eval': \strong{\verb{numeric vector}} of double containing the evaluation of the beta along a grid.
\item 'basis_coeff': \strong{\verb{numeric vector}} of double containing the grid (element abscissa_ev).
}
}
}
\description{
Function to perform predictions on new statistical units using a fitted Functional Mixed Geographically Weighted Regression model. Non-stationary betas have to be recomputed in the new locations.
}
\details{
Covariates of units to be predicted have to be sampled in the same sample points for which the training data have been (t_points of FMSGWR_ESC).
Covariates basis expansion for the units to be predicted has to be done with respect to the basis used for the covariates in the training set
}
\references{
\itemize{
\item Source code: \href{https://github.com/AndreaEnricoFranzoni/FunctionalMultiSourceGeographicallyWeightedRegression}{fdagwr implementation}
}
}
\author{
Andrea Enrico Franzoni
}
